Adding the preprocessed data as well as the two functions necessary to clean the paris geodf and map the information
