--- ohsome_api/land_coverage.py
@@ -1,4 +1,0 @@
-
-
-
-def clean_data(df)
